requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.4.0=pypi_0
aiofiles=22.1.0=pypi_0
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
aiosqlite=0.18.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.6.2=pypi_0
appnope=0.1.3=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=22.2.0=pypi_0
babel=2.11.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.2=pypi_0
binutils_impl_linux-64=2.40=hf600244_0
bleach=6.0.0=pypi_0
brotli-python=1.1.0=py38h17151c0_1
bzip2=1.0.8=hd590300_5
ca-certificates=2024.7.4=hbcca054_0
cached-property=1.5.2=pypi_0
cachetools=5.3.2=pypi_0
calvin-env=0.0.1=pypi_0
certifi=2022.12.7=pypi_0
cffi=1.15.1=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.0.1=pypi_0
chex=0.1.6=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
cmake=3.18.4=pypi_0
colorama=0.4.6=pypi_0
colorlog=6.8.0=pypi_0
comm=0.1.2=pypi_0
commonmark=0.9.1=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.0.7=pypi_0
cudatoolkit=11.8.0=h4ba93d1_12
cycler=0.11.0=pypi_0
cython=0.29.33=pypi_0
d4rl=1.1=pypi_0
debugpy=1.6.6=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
distrax=0.1.3=pypi_0
dm-control=1.0.10=pypi_0
dm-env=1.6=pypi_0
dm-tree=0.1.8=pypi_0
docker-pycreds=0.4.0=pypi_0
et_xmlfile=1.1.0=pyhd8ed1ab_0
etils=1.0.0=pypi_0
executing=1.2.0=pypi_0
faiss=1.7.4=py38cuda112h48d0473_0_cuda
faiss-gpu=1.7.4=h788eb59_0
farama-notifications=0.0.4=pypi_0
fasteners=0.18=pypi_0
fastjsonschema=2.16.3=pypi_0
filelock=3.12.4=pypi_0
flatbuffers=23.5.26=pypi_0
flax=0.6.4=pypi_0
fonttools=4.38.0=pypi_0
fqdn=1.5.1=pypi_0
freetype-py=2.4.0=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.4.0=pypi_0
future=0.18.3=pypi_0
gast=0.5.3=pypi_0
gcc=13.2.0=h574f8da_2
gcc_impl_linux-64=13.2.0=h338b0a0_3
gdown=4.7.1=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.30=pypi_0
glcontext=2.5.0=pypi_0
glfw=1.12.0=pypi_0
google-auth=2.23.4=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.59.3=pypi_0
gym=0.23.1=pypi_0
gym-notices=0.0.8=pypi_0
gym3=0.3.3=pypi_0
gymnasium=0.29.1=pypi_0
h5py=3.8.0=pypi_0
huggingface-hub=0.17.3=pypi_0
hydra-colorlog=1.2.0=pypi_0
hydra-core=1.3.2=pypi_0
icecream=2.1.3=pypi_0
idna=3.4=pypi_0
imageio=2.25.1=pypi_0
imageio-ffmpeg=0.3.0=pypi_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.10.2=pypi_0
intel-openmp=2024.0.0=pypi_0
ipykernel=6.21.2=pypi_0
ipython=8.10.0=pypi_0
ipython-genutils=0.2.0=pypi_0
isoduration=20.11.0=pypi_0
jax=0.4.3=pypi_0
jaxlib=0.4.3+cuda11.cudnn82=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
json5=0.9.11=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter-client=8.0.3=pypi_0
jupyter-core=5.2.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.3.0=pypi_0
jupyter-server-fileid=0.8.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyter-server-ydoc=0.6.1=pypi_0
jupyter-ydoc=0.2.2=pypi_0
jupyterlab=3.6.1=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.19.0=pypi_0
jupyterlab-vim=0.15.1=pypi_0
keras=2.12.0=pypi_0
kernel-headers_linux-64=2.6.32=he073ed8_16
kiwisolver=1.4.4=pypi_0
labmaze=1.0.6=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libblas=3.9.0=20_linux64_openblas
libcblas=3.9.0=20_linux64_openblas
libclang=16.0.6=pypi_0
libfaiss=1.7.4=cuda112hb18a002_0_cuda
libfaiss-avx2=1.7.4=cuda112h1234567_0_cuda
libffi=3.4.2=h7f98852_5
libgcc-devel_linux-64=13.2.0=ha9c7c90_103
libgcc-ng=13.2.0=h807b86a_3
libgfortran-ng=13.2.0=h69a702a_3
libgfortran5=13.2.0=ha4646dd_3
libgomp=13.2.0=h807b86a_3
liblapack=3.9.0=20_linux64_openblas
libnsl=2.0.1=hd590300_0
libopenblas=0.3.25=pthreads_h413a1c8_0
libsanitizer=13.2.0=h7e041cc_3
libsqlite=3.44.0=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_3
libuuid=2.38.1=h0b41bf4_0
libzlib=1.2.13=hd590300_5
lightning-utilities=0.10.0=pypi_0
llvmlite=0.41.1=pypi_0
lxml=4.9.2=pypi_0
markdown=3.5.1=pypi_0
markdown-it-py=2.1.0=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.4=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mistune=2.0.5=pypi_0
mjrl=1.0.0=pypi_0
mkl=2024.0.0=pypi_0
ml-collections=0.1.1=pypi_0
moderngl=5.10.0=pypi_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.4=pypi_0
mujoco=2.3.2=pypi_0
mujoco-py=2.1.2.14=pypi_0
multicoretsne=0.1=pypi_0
multidict=6.0.4=pypi_0
nbclassic=0.5.2=pypi_0
nbclient=0.7.2=pypi_0
nbconvert=7.2.9=pypi_0
nbformat=5.7.3=pypi_0
ncurses=6.4=h59595ed_2
nest-asyncio=1.5.6=pypi_0
networkx=3.0=pypi_0
nltk=3.8.1=pypi_0
notebook=6.5.2=pypi_0
notebook-shim=0.2.2=pypi_0
numba=0.58.1=pypi_0
numpy=1.22.4=pypi_0
numpy-quaternion=2022.4.4=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.8.1.78=pypi_0
openpyxl=3.1.2=py38h01eb140_1
openssl=3.3.1=h4bc722e_2
opentsne=1.0.1=py38h3135665_2
opt-einsum=3.3.0=pypi_0
optax=0.1.4=pypi_0
orbax=0.1.1=pypi_0
packaging=23.0=pypi_0
pandas=2.0.3=py38h01efb38_1
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=pypi_0
pip=23.3.1=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.0.0=pypi_0
plotly=5.18.0=pypi_0
ply=3.11=pypi_0
pooch=1.7.0=pypi_0
procgen=0.10.7=pypi_0
proglog=0.1.10=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.37=pypi_0
protobuf=4.21.12=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pybullet=3.2.5=pypi_0
pycollada=0.6=pypi_0
pycparser=2.21=pypi_0
pydeprecate=0.3.1=pypi_0
pyglet=1.5.0=pypi_0
pygments=2.14.0=pypi_0
pynndescent=0.5.11=pypi_0
pyopengl=3.1.6=pypi_0
pyopengl-accelerate=3.1.7=pypi_0
pyparsing=3.0.9=pypi_0
pyproject-toml=0.0.10=pypi_0
pyqt5=5.15.9=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.12.2=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.18=hd12c33a_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-json-logger=2.0.7=pypi_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.8=4_cp38
pytorch-lightning=1.4.9=pypi_0
pytz=2022.7.1=pypi_0
pyvista=0.39.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.0=pypi_0
readline=8.2=h8228510_1
regex=2023.10.3=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=11.1.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.1=pypi_0
scikit-learn=1.3.2=py38ha25d942_2
scipy=1.10.0=pypi_0
scooby=0.7.4=pypi_0
send2trash=1.8.0=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=1.15.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=57.5.0=pypi_0
simpy=4.0.2=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4=pypi_0
stack-data=0.6.2=pypi_0
sympy=1.12=pypi_0
sysroot_linux-64=2.12=he073ed8_16
tbb=2021.11.0=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.12.3=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.12.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
tensorflow-probability=0.19.0=pypi_0
tensorstore=0.1.32=pypi_0
termcolor=2.2.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.15.0=pypi_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.0=pypi_0
torch=2.1.1+cu118=pypi_0
torchaudio=2.1.1+cu118=pypi_0
torchmetrics=1.2.1=pypi_0
torchvision=0.16.1+cu118=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.35.2=pypi_0
trimesh=4.0.5=pypi_0
triton=2.1.0=pypi_0
tsnecuda=3.0.1+cu118=pypi_0
typing-extensions=4.5.0=pypi_0
umap-learn=0.5.5=pypi_0
uri-template=1.2.0=pypi_0
urllib3=1.26.14=pypi_0
wandb=0.13.10=pypi_0
wcwidth=0.2.6=pypi_0
webcolors=1.12=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.3=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
xxhash=3.4.1=pypi_0
xz=5.2.6=h166bdaf_0
y-py=0.5.9=pypi_0
yarl=1.9.3=pypi_0
ypy-websocket=0.8.2=pypi_0
zipp=3.13.0=pypi_0
